feat(CORE/Script Hooks): Alternative proc reactions and alternative attack power calcs
The purpose of this change is to add script hooks to support the creation of a module allowing for a classless wow experience.

The first of these script hooks is to allow an alternative method to be used to calculate a player's attack power and damage.

The second of these script hooks is to allow an alternative method to be used to set reactions to procs- for the purpose of reactive defense.

The last change included in this is changing a val2 variable name to something more useful, like baseAttackPower in StatSystem->UpdateAttackPowerAndDamage